update: seed data with sqlfile
