+++ pkg/infra/dao/testdata/delete.sql
@@ -1,0 +1,2 @@
+-- write all tables to use as seed
+delete from invitations;
