+++ pkg/infra/dao/testdata/invitation/find_success.sql
@@ -1,0 +1,5 @@
+insert into invitations (
+  id, user_id, code, created_at, updated_at
+) values (
+  1, 1, 'code', now(), now()
+);
