--- pkg/infra/dao/testdata/invitation/invitations.yml
@@ -1,5 +1,0 @@
-- id: 1
-  user_id: 1
-  code: code
-  created_at: 2020-01-01 00:00:00
-  updated_at: 2020-01-01 00:00:00
